Merge branch 'master' of https://github.com/ligerbots/RecycleRush2015Robot into moreDriveWork

Conflicts:
	src/Commands/DriveDistance.h
	src/Commands/DriveDistanceCommand.cpp

+++ src/Commands/DriveCommand.h
@@ -30,7 +30,7 @@ public:
 	virtual void End();
 	virtual void Interrupted();
 private:
-	float turnMax;
+	double turnMax;
 };
 
 #endif
